Crate re_component_ui

This crate implements various component editors.

The only entry point is create_component_ui_registry, which registers all editors in the component UI registry. This should be called by re_viewer on startup.

Modules

color ğŸ”’

datatype_uis ğŸ”’

entity_path ğŸ”’

fallback_ui ğŸ”’

image_format ğŸ”’

line_strip ğŸ”’

marker_shape ğŸ”’

pinhole ğŸ”’

radius ğŸ”’

resolution ğŸ”’

response_utils ğŸ”’

timeline ğŸ”’

transforms ğŸ”’

video_timestamp ğŸ”’

view_coordinates ğŸ”’

visual_bounds2d ğŸ”’

Functions

create_component_ui_registry

Crates a component ui registry and registers all editors of this crate to it.