Trait re_component_ui::datatype_uis::enum_combobox::VariantAvailableProvider

pub trait VariantAvailableProvider<EnumT: Enum> {
    // Required method
    fn is_variant_enabled(
        ctx: &ViewerContext<'_>,
        variant: EnumT,
    ) -> VariantAvailable;
}

Trait for a type that can provide information about whether a particular variant of an enum is available for selection.

Required Methods

fn is_variant_enabled( ctx: &ViewerContext<'_>, variant: EnumT, ) -> VariantAvailable

Object Safety

This trait is not object safe.