Crate re_component_ui

This crate implements various component editors.

The only entry point is create_component_ui_registry, which registers all editors in the component UI registry. This should be called by re_viewer on startup.

Modules

color ğŸ”’

datatype_uis ğŸ”’

entity_path ğŸ”’

fallback_ui ğŸ”’

geo_line_string ğŸ”’

image_format ğŸ”’

lat_lon ğŸ”’

line_strip ğŸ”’

map_provider ğŸ”’

marker_shape ğŸ”’

pinhole ğŸ”’

plane3d ğŸ”’

radius ğŸ”’

resolution ğŸ”’

response_utils ğŸ”’

timeline ğŸ”’

transforms ğŸ”’

video_timestamp ğŸ”’

view_coordinates ğŸ”’

visual_bounds2d ğŸ”’

zoom_level ğŸ”’

Constants

DEFAULT_NUMBER_WIDTH ğŸ”’

Default number of ui points to show a number.

Functions

create_component_ui_registry

Crates a component ui registry and registers all editors of this crate to it.